refactor(core): move from jsdoc to ts (again)
