+++ packages/core/package.json
@@ -1,32 +1,28 @@
 {
   "name": "@briddge/core",
-  "version": "0.3.13",
+  "version": "0.3.19",
   "packageManager": "pnpm@8.15.3",
   "type": "module",
-  "main": "./dist/briddge-core.cjs.js",
-  "module": "./dist/briddge-core.esm-bundler.js",
-  "types": "./src/index.d.ts",
   "scripts": {
     "test": "vitest --",
-    "build": "pnpm build-rm-alias && pnpm build-js",
-    "build-js": "rimraf dist && pnpm rollup -c rollup.config.js",
-    "build-vue": "(rimraf vue/temp && cd vue && pnpm tsc --noCheck) || exit 0;",
-    "build-rm-alias": "(pnpm build-vue && pnpm dlx tsc-alias -p vue/tsconfig.json) || exit 0;",
+    "build": "pnpm build-rm-alias && pnpm build-js && pnpm build-types",
+    "build-js": "rimraf dist && rollup -c rollup.config.js",
+    "build-types": "tsc --noCheck",
+    "build-vue": "pnpm build-vue-js && pnpm build-rm-vue-aliases",
+    "build-vue-js": "(rimraf vue/temp && cd vue && pnpm tsc --noCheck) || exit 0;",
+    "build-rm-vue-aliases": "(pnpm dlx tsc-alias -p vue/tsconfig.json) || exit 0;",
     "release": "release-it"
   },
   "exports": {
     ".": {
-      "import": {
-        "default": "./dist/briddge-core.esm-bundler.js",
-        "types": "./src/index.d.ts"
-      },
-      "require": {
-        "default": "./dist/briddge-core.cjs.js",
-        "types": "./src/index.d.ts"
-      }
+      "import": "./dist/index.js",
+      "require": "./dist/index.cjs.js"
     },
-    "./package.json": "./package.json"
+    "./*": "./*"
   },
+  "main": "dist/index.cjs.js",
+  "module": "dist/index.js",
+  "types": "types/index.d.ts",
   "imports": {
     "#vue-internals/reactivity/*.js": "./vue/temp/packages/reactivity/src/*.js",
     "#vue-internals/runtime-core/*.js": "./vue/temp/packages/runtime-core/src/*.js",
@@ -38,11 +34,7 @@
   "sideEffects": false,
   "files": [
     "dist",
-    "src",
-    "vue/temp/packages/runtime-core/src",
-    "vue/temp/packages/reactivity/src",
-    "vue/temp/packages/shared/src",
-    "global.d.ts"
+    "types"
   ],
   "peerDependencies": {
     "react": ">=18.0.0"
@@ -61,6 +53,7 @@
     "rollup-plugin-dts": "^6.1.1",
     "rollup-plugin-esbuild": "^6.1.1",
     "tsc-alias": "^1.8.10",
+    "tsup": "^8.3.5",
     "typescript": "^5.7.2",
     "unplugin-inline-enum": "^0.2.0",
     "vite": "^6.0.1"
