+++ packages/core/tsconfig.json
@@ -1,49 +1,45 @@
 {
-  "compilerOptions": {
-    "baseUrl": ".",
-    "target": "ES2020",
-    "useDefineForClassFields": true,
-    "module": "ESNext",
-    "lib": ["ES2020", "DOM", "DOM.Iterable"],
-    "sourceMap": false,
-    "moduleResolution": "bundler",
-    "allowJs": true,
-    "strict": true,
-    "noUnusedLocals": true,
-    "experimentalDecorators": true,
-    "resolveJsonModule": true,
-    "isolatedModules": true,
-    "skipLibCheck": true,
-    "esModuleInterop": true,
-    "removeComments": false,
-    "jsx": "react-jsx",
-    "rootDir": ".",
-    "paths": {
-      "#vue-internals/reactivity/*": ["vue/packages/reactivity/src/*"],
-      "#vue-internals/runtime-core/*": ["vue/packages/runtime-core/src/*"],
-      "#vue-internals/shared/*": ["vue/packages/shared/src/*"],
-      "#vue/reactivity": ["vue/packages/reactivity/src"],
-      "#vue/runtime-core": ["vue/packages/runtime-core/src"],
-      "#vue/shared": ["vue/packages/shared/src"]
+    "compilerOptions": {
+        "baseUrl": ".",
+        "target": "ES2022",
+        "useDefineForClassFields": true,
+        "outDir": "dist/types",
+        "module": "ESNext",
+        "lib": ["ES2020", "DOM", "DOM.Iterable"],
+        "sourceMap": false,
+        "emitDeclarationOnly": true,
+        "moduleResolution": "bundler",
+        "allowJs": true,
+        "noCheck": true,
+        "strict": true,
+        "noUnusedLocals": true,
+        "experimentalDecorators": true,
+        "resolveJsonModule": true,
+        "isolatedModules": true,
+        "skipLibCheck": true,
+        // "esModuleInterop": true,
+        "removeComments": false,
+        "jsx": "react-jsx",
+        "rootDir": ".",
+        "paths": {
+            "#vue-internals/reactivity/*": ["vue/packages/reactivity/src/*"],
+            "#vue-internals/runtime-core/*": ["vue/packages/runtime-core/src/*"],
+            "#vue-internals/shared/*": ["vue/packages/shared/src/*"],
+            "#vue/reactivity": ["vue/packages/reactivity/src"],
+            "#vue/runtime-core": ["vue/packages/runtime-core/src"],
+            "#vue/shared": ["vue/packages/shared/src"],
+            "@vue/reactivity": ["vue/packages/reactivity/src"],
+            "@vue/runtime-core": ["vue/packages/runtime-core/src"],
+            "@vue/shared": ["vue/packages/shared/src"]
+        },
+        "types": ["vitest/globals", "node"],
+        "composite": true
     },
-    "types": ["vitest/globals", "node"],
-    "composite": true
-  },
-  "include": [
-    "src",
-    "vue/packages/*/src",
-    "vue/packages/*/__tests__",
-    "vue/packages/global.d.ts",
-    "vue/packages/*/src",
-    "vue/packages/*/__tests__",
-    "vue/packages/vue/jsx-runtime",
-    "vue/packages/runtime-dom/types/jsx.d.ts",
-    "vue/packages/vue/react-helpers",
-    "vue/packages/vue/__benchmarks__/apiWatch.bench.ts",
-    "vue/packages/vue/__benchmarks__/computed.bench.ts",
-    "vue/packages/vue/__benchmarks__/reactiveArray.bench.ts",
-    "vue/packages/vue/__benchmarks__/reactiveMap.bench.ts",
-    "vue/packages/vue/__benchmarks__/reactiveObject.bench.ts",
-    "vue/packages/vue/__benchmarks__/ref.bench.ts"
-,  ]
+    "include": [
+        "src",
+        "vue/packages/reactivity/src",
+        "vue/packages/runtime-core/src",
+        "vue/packages/shared/src",
+        "vue/packages/global.d.ts"
+    ]
 }
